Move QR into separate module
